move resource files. rename External folder to ext for consistant naming.
