--- api/Framework.h
@@ -1,7 +1,0 @@
-#pragma once
-
-//#define WIN32_LEAN_AND_MEAN             // Exclude rarely-used stuff from Windows headers
-
-// Header files
-
-
