+++ api/ext/DLL.rc
