+++ api/ext/obfuscate.h
