+++ api/ext/resource.h
