+++ src/ServerInterface.cpp
@@ -1,7 +1,7 @@
 #include "ServerInterface.h"
 #include "Serialization.h"
 #include "NetworkManager.h"
-#include "External/base64.h"
+#include "ext/base64.h"
 #include "Logging.hpp"
 
 #include <iostream>
